requirements.txt
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
PyYAML==6.0.2
feedparser==6.0.11
vaderSentiment==3.3.2
yfinance==1.2.0
nsepython==2.97

streamlit==1.41.1
